+++ user/src/main/java/com/team15gijo/user/domain/service/UserDomainService.java
@@ -1,9 +1,11 @@
 package com.team15gijo.user.domain.service;
 
 import com.team15gijo.user.domain.model.UserEntity;
+import com.team15gijo.user.domain.model.UserRegionEntity;
 import com.team15gijo.user.presentation.dto.request.v1.UserSignUpRequestDto;
 
 public interface UserDomainService {
 
-    UserEntity createUser(UserSignUpRequestDto userSignUpRequestDto);
+    UserEntity createUser(UserSignUpRequestDto userSignUpRequestDto,
+            UserRegionEntity userRegionEntity);
 }
